Seed, styling and display
Most important is creating the seed.py file.
This allow to quickly populate the database with dummy blogs.
- Created with the help from chatgpt.

Second i added some slight css in the style.css file.   I tested out with an extreme background-color: red; to see the connection which it did.

I made sure that the base html accepted this css in the head.

And just adjusted home.html abit.